Beginnings of a really large refactor.
Building all of our Cogs out of a base class where we can store some common bits.

Breaking the pretty huge puzzles cog into more manageable pieces (especially since some of the pieces have nothing to do with puzzles at all?)

First steps towards abstracting away all of the interaction with the database into a more manageable controller chunk, so we can more easily handle some of the database moves I plan to do in the near-ish future.
