+++ .gitignore
@@ -135,5 +135,5 @@ dmypy.json
 # TEMPLATE END
 config.json
 *_secrets.json
-data/*
 /.vs
+
