--- bot/cogs/cog.template
@@ -1,15 +1,0 @@
-import discord
-from discord.ext import commands
-
-
-class Name(commands.Cog):
-    def __init__(self, bot):
-        self.bot = bot
-
-    @commands.Cog.listener()
-    async def on_ready(self):
-        print(f'{type(self).__name__} Cog ready.')
-
-
-def setup(bot):
-    bot.add_cog(Name(bot))
